move components to new dir
