+++ frontend/packs/components/others/GetCenter.vue
@@ -11,7 +11,7 @@
 
 <script>
 
-import RoomForm from './RoomForm.vue'
+import RoomForm from '../form/RoomForm.vue'
 
 export default {
